Count dynamic basic blocks (streams) in pintool
streamcount.cpp was previously counting static basic blocks.

Pin BBLs are static, and defined as a contiguous sequence of
instructions that starts at the target of a taken branch and ends
with a branch instruction (regardless of whether it is taken or not.)

Dynamic BBLs (or streams) are defined as a contiguous sequence of
instructions that starts at the target of a taken branch and ends
with the first TAKEN branch in the sequence. Thus, one stream can
encompass multiple basic blocks.